Move test init out of the gallery image tests
